Legacy + Scary Items options
Ability to choose to not get Legacy Skins and also not "Scary" Cosmetics, that are completely exclusive.

Missing dev bnr and bdg, will find them later and add, can just be added to the ScaryItems List in Extras

Also fixed incorrect error message in CDN and also now showing which file its trying to generate, so if it times out, you can see it begins to generate killswitch, before it looked like the program has just died.